Update and rename gradient_finder.c to gradient_finder.hapcomments.0131201.c
So this looks OK. At least I can read the code, which is very good indeed for anything written in C.  I made some comments regarding the use of #defines and some other stuff to increase the speed of the code.
More important, I think the transition when phi becomes greater than pi doesn't follow the conservation of momentum. I wrote a bit about this in the code, and now that I have a chance to think about it, consider a somewhat obsessive explored who is going to walk and swim from pole to pole. However, being obsessive the path is chosen to always be in the direction of Greenwich. So the expedition starts at the North pole, walks South in the direction of Greenwich (in the Western hemisphere) and thence to the South Pole. At the south pole they keep walking, but having gone as far south as possible they are now walking North (i.e. in the opposite phi direction) and have crossed over into the Eastern hemisphere ( i.e. Pi radians advanced from their previous course).

If you were to plan such an expedition, I don't recommend this route. It is a very long swim from the South pole to the North and there is nowhere good to stop for dinner, at least until you get to Russia. And beets and potatoes make a poor dinner after a long swim.