Update and rename gradient_finder.c to gradient_finderHAPcomments3.c
So this needs the following done now:
Better documentation of subroutines
Convert the array mess into a structure,
Write a general routine to check for an correct the theta and phi related values, both position and velocities.

I haven't read through the file thing at the end yet. I'll do that and see if I have anything creative to say.